feat: Add pyinstaller.bat file for building executable with GUI image support
A new pyinstaller.bat file has been added to handle the building of the executable file. This file includes a TODO comment to address the issue of GUI image handling and specifies an icon for the exe file. The command "pyinstaller --onefile --noconsole main.py" is used to build the executable.

Note: Please remove any meta information such as issue references, tags, or author names from the commit message.